+++ packages/core/src/types.ts
@@ -4,5 +4,6 @@ export * from "./types/tools.js";
 export * from "./types/options.js";
 export * from "./types/meta.js";
 export * from "./types/session.js";
+export * from "./types/requests.js";
 export * from "./types/responses.js";
 export * from "./types/sse.js";